backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
distlib==0.4.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
filelock==3.20.3
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pipenv==2026.0.3
platformdirs==4.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==80.10.2
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
virtualenv==20.36.1
